Add set to null if invalid prop is set.
If a user puts in an invalid youtube prop it resets it to null and doesnt allow for them to submit invalid props.  This should fix the `glitch`, but not that if the company currenty has props and the enter new invalid ones it will attempt to set it to null and the editor will have to discard the changes.